requirements.txt
flask==3.0.3
pymupdf==1.24.9
python-docx==1.1.2
transformers==4.44.2
torch>=2.2.0
requests>=2.31.0
